requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
pydantic==2.5.0

# P0.2: RSA key management for RS256 JWT signing
cryptography==41.0.7

# unison-common package (local install)
git+https://github.com/project-unisonOS/unison-common.git@main
